Area Based Sound Echo Improvement: Patch 1 (#56)
* major refactor

not done, needs commenting and adjustments and namespace changes
probably

* refactored the raycasting system. instead of tilemarching, we use
robust engine's rays with predicate. rays also bounce and come off of
the player instead of audio sources (achieves the same effect for much
cheaper)

* refactored the way echo is calculated. rays will keep track of
entities that have the new client-side component
AudioAbsorptionComponent (probably gonna rename it to like...
AudioAcousticsComponent or something). averages out all the absorption
values it finds and does coefficient stuff and i still need to tweak
this a bit

* ignore the blood stains

* fixed up echo calculations

* converted audio absorption to 0~100% yaml wise

* removed unused parameters/usings

* fix collision groups

* re-added roof and space check

* comment out logging

* oops

i messed up with typos but fixed it